app/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9
pydantic>=2.6
starlette>=0.36
speechify-api>=0.19.1
python-dotenv>=1.0